refactor: test folder structure
